+++ src/Modules/Store/DTO/StoreStatusForMember.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Foodsharing\Modules\Store\DTO;
+
+class StoreStatusForMember
+{
+	public MinimalStoreIdentifier $store;
+	public bool $isManaging;
+	public int $membershipStatus;
+	public ?int $pickupStatus;
+}