+++ src/Modules/Store/StoreGateway.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Foodsharing\Modules\Store;
 
 use Foodsharing\Modules\Core\BaseGateway;
@@ -10,7 +12,8 @@ use Foodsharing\Modules\Core\DBConstants\StoreTeam\MembershipStatus;
 use Foodsharing\Modules\Region\RegionGateway;
 use Foodsharing\Modules\Store\DTO\CreateStoreData;
 use Foodsharing\Modules\Store\DTO\Store;
-use Foodsharing\Modules\Store\DTO\StoreForTopbarMenu;
+use Foodsharing\Modules\Store\DTO\StoreTeamMembership;
+use UnexpectedValueException;
 
 class StoreGateway extends BaseGateway
 {
@@ -763,40 +766,49 @@ class StoreGateway extends BaseGateway
 	}
 
 	/**
-	 * @return StoreForTopbarMenu[]
+	 * Returns a list with all store memberships of the foodsaver.
+	 *
+	 * @param int $fsId Foodsharer Id
+	 * @param int[] $storeCooperationStates All store state should should be contained @see CooperationStatus
+	 *
+	 * @return StoreTeamMembership[] Returns a array of memberships
 	 */
-	public function listFilteredStoresForFoodsaver($fsId): array
+	public function listAllStoreTeamMembershipsForFoodsaver(int $fsId, array $storeCooperationStates)
 	{
-		$rows = $this->db->fetchAll('
-			SELECT 	b.`id`,
-					b.name,
-					bt.verantwortlich AS managing
-
-			FROM 	`fs_betrieb_team` bt
-					INNER JOIN `fs_betrieb` b
-			        ON bt.betrieb_id = b.id
-
-			WHERE   bt.`foodsaver_id` = :fsId
-			AND 	bt.active = :membershipStatus
-			AND 	b.betrieb_status_id NOT IN (:doesNotWantToWorkWithUs, :givesToOtherCharity)
-			ORDER BY bt.verantwortlich DESC, b.name ASC
-		', [
-			':fsId' => $fsId,
-			':membershipStatus' => MembershipStatus::MEMBER,
-			':doesNotWantToWorkWithUs' => CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US,
-			':givesToOtherCharity' => CooperationStatus::GIVES_TO_OTHER_CHARITY
-		]);
-
-		$stores = [];
-		foreach ($rows as $row) {
-			$store = new StoreForTopbarMenu();
-			$store->id = $row['id'];
-			$store->name = $row['name'];
-			$store->isManaging = $row['managing'];
-			$stores[] = $store;
+		if ($fsId == 0) {
+			return [];
 		}
 
-		return $stores;
+		// last check of CooperationStatus content before DB
+		foreach ($storeCooperationStates as $storeState) {
+			if (!CooperationStatus::isValidStatus($storeState)) {
+				throw new UnexpectedValueException('Store cooperation state is not valid.');
+			}
+		}
+
+		$inPlaceHolder = implode(', ', array_fill(0, count($storeCooperationStates), '?'));
+		$rows = $this->db->fetchAll('
+			SELECT 	b.id as store_id,
+					b.name as store_name,
+					bt.verantwortlich AS managing,
+					bt.active as membership_status
+			FROM fs_betrieb_team bt
+				INNER JOIN fs_betrieb b
+					ON bt.betrieb_id = b.id
+			WHERE   bt.`foodsaver_id` = ?
+			AND 	b.betrieb_status_id IN (' . $inPlaceHolder . ')
+			ORDER BY bt.verantwortlich DESC, membership_status ASC, b.name ASC
+		', [
+			$fsId,
+			$storeCooperationStates
+		]);
+
+		$results = [];
+		foreach ($rows as $row) {
+			$results[] = StoreTeamMembership::createFromArray($row);
+		}
+
+		return $results;
 	}
 
 	public function listStoreIds($fsId)
@@ -913,8 +925,8 @@ class StoreGateway extends BaseGateway
 			'fs_id_a' => $foodsaver_id,
 			'fs_id_p' => $fs_id_p,
 			'date_reference' => $dateReference ? $dateReference->format('Y-m-d H:i:s') : null,
-			'content' => strip_tags($content),
-			'reason' => strip_tags($reason),
+			'content' => $content ? strip_tags($content) : '',
+			'reason' => $reason ? strip_tags($reason) : ''
 		]);
 	}